requirements.txt
Flask==1.1.2
Flask-Script==2.0.6
celery==4.4.7
flower==0.9.5
Flask-Mail==0.9.1
boto3==1.14.47
gunicorn==20.0.4
flask-jwt-extended==3.24.1
mongoengine==0.20.0
botocore==1.17.47
Werkzeug~=1.0.1
flask-cors==3.0.8
bcrypt==3.2.0
pymongo~=3.11.0
Coverage==5.2.1
PyJWT~=1.7.1
redis==3.5.3
requests==2.24.0
